Refactor linear solver algorithm tests into helpers
