+++ tests/Unit/NumericalAlgorithms/LinearSolver/ConjugateGradient/Test_ConjugateGradientAlgorithm.input
@@ -1,4 +1,9 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+LinearOperator: [[4, 1], [1, 3]]
+Source: [1, 2]
+InitialGuess: [2, 1]
+ExpectedResult: [0.0909090909090909, 0.6363636363636364]
+
 Verbosity: Verbose
