+++ tests/Unit/NumericalAlgorithms/LinearSolver/ConjugateGradient/Test_DistributedConjugateGradientAlgorithm.input
@@ -1,4 +1,28 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+NumberOfElements: 2
+
+LinearOperator:
+    - [[ 5.305164769729845,  0.848826363156775, -0.742723067762178],
+       [ 0.848826363156775,  3.395305452627101, -0.424413181578388],
+       [-0.742723067762178, -0.424413181578388,  3.395305452627101],
+       [ 0.318309886183791, -1.273239544735163, -1.909859317102744],
+       [ 0.               ,  0.               , -1.273239544735163],
+       [ 0.               ,  0.               ,  0.318309886183791]]
+    - [[ 0.318309886183791,  0.               ,  0.               ],
+       [-1.273239544735163,  0.               ,  0.               ],
+       [-1.909859317102744, -1.273239544735163,  0.318309886183791],
+       [ 3.395305452627101, -0.424413181578388, -0.742723067762178],
+       [-0.424413181578388,  3.395305452627101,  0.848826363156775],
+       [-0.742723067762178,  0.848826363156775,  5.305164769729845]]
+
+Source:
+    - [0.                , 0.740480489693061, 0.2617993877991494]
+    - [0.2617993877991494, 0.740480489693061, 0.                ]
+
+ExpectedResult:
+    - [-0.0363482510397858,  0.7235793356729757,  0.9928055333486293]
+    - [ 0.9928055333486292,  0.7235793356729758, -0.0363482510397858]
+
 Verbosity: Verbose
